% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{get_test_query}
\alias{get_test_query}
\title{Get Test Query}
\usage{
get_test_query(endpoint)
}
\arguments{
\item{endpoint}{The web service resource you want a test query for. \code{endpoint}
must be one of the following: "patents", "inventors", "assignees",
"locations", "cpc_groups", "cpc_subgroups", "cpc_subsections", "uspc_mainclasses",
"uspc_subclasses","nber_categories", "nber_subcategories", "application_citations",
or "patent_citations"}
}
\value{
a test query for the specified endpoint.
}
\description{
In the new version of the api, only three of the endpoints are searchable
by patent number. This function provides a sample query for each
endpoint, except for locations, which isn't on the test server yet
}
\examples{
\dontrun{

get_test_query("patents")
}

}
