% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-pv.R
\name{pad_patent_id}
\alias{pad_patent_id}
\title{Pad patent_id}
\usage{
pad_patent_id(patent_id)
}
\arguments{
\item{patent_id}{The patent_id that needs to be padded.  It can
be the patent_id for a utility, design, plant or reissue patent.}
}
\description{
This function strategically pads a patent_id with zeroes to 8 characters,
needed only for custom paging that uses sorts by patent_id.
}
\examples{
\dontrun{
padded <- pad_patent_id("RE36479")

padded2 <- pad_patent_id("3930306")
}

}
