% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patentsview-package.r
\docType{package}
\name{patentsview}
\alias{patentsview}
\alias{patentsview-package}
\title{patentsview: An R client to the PatentsView API}
\description{
patentsview: An R client to the PatentsView API
}
\section{About}{

}

\section{Documentation}{

}

