% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-dsl.R
\docType{data}
\name{qry_funs}
\alias{qry_funs}
\title{Query function list}
\format{An object of class \code{list} of length 14.}
\usage{
qry_funs
}
\description{
A list containing R functions that make it easy to use the elastic-like query operators used by the PatentsView API to write a query.
}
\keyword{datasets}

