% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fieldsdf}
\alias{fieldsdf}
\title{Fields data frame}
\format{
A data frame with the following columns:
\describe{
  \item{endpoint}{The endpoint that this field record is for}
  \item{field}{The complete name of the field, including the parent group if
  applicable}
  \item{data_type}{The field's input data type}
  \item{group}{The group the field belongs to}
  \item{common_name}{The field name without the parent group structure}
}
}
\usage{
fieldsdf
}
\description{
A data frame containing the names of retrievable fields for each of the
endpoints. You can find this data on the API's online documentation for each
endpoint as well (e.g., the
\href{https://patentsview.org/apis/api-endpoints/patents}{patents endpoint
field list table}).
}
\keyword{datasets}
