% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fields.R
\name{get_endpoints}
\alias{get_endpoints}
\title{Get endpoints}
\usage{
get_endpoints()
}
\value{
A character vector with the names of the 9 endpoints. Those endpoints are:

\itemize{
   \item assignee
   \item cpc_subsection
   \item cpc_group
   \item cpc_subgroup
   \item inventors
   \item locations
   \item nber_category
   \item nber_subcategory
   \item patents
   \item uspc_mainclass
   \item uspc_subclass
   \item application_citation
   \item patent_citation
 }
}
\description{
This function reminds the user what the 9 possible PatentsView API endpoints
are.
}
\examples{
get_endpoints()
}
