% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fields.R
\name{get_endpoints}
\alias{get_endpoints}
\title{Get endpoints}
\usage{
get_endpoints()
}
\value{
A character vector with the names of each endpoint.
}
\description{
This function reminds the user what the possible PatentSearch API endpoints
are.  (Note that the API was originally know as the PatentsView API.)
}
