% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-fields.R
\name{get_endpoints}
\alias{get_endpoints}
\title{Get endpoints}
\usage{
get_endpoints()
}
\value{
A character vector with the names of the 7 endpoints. Those endpoints are:

\itemize{
   \item assignees
   \item cpc_subsections
   \item inventors
   \item locations
   \item nber_subcategories
   \item patents
   \item uspc_mainclasses
 }
}
\description{
This function reminds the user what the 7 possible PatentsView API endpoints
are.
}
\examples{
get_endpoints()
}
