% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-attribute-helpers.R
\name{get_ok_pk}
\alias{get_ok_pk}
\title{Get OK primary key}
\usage{
get_ok_pk(endpoint)
}
\arguments{
\item{endpoint}{The endpoint which you would like to know a potential primary
key for.}
}
\value{
The column names that represent a single row for the given endpoint.
}
\description{
This function suggests column(s) that you could use for the \code{pk} argument
in \code{\link{unnest_pv_data}}, based on the endpoint you searched.
It will return a potential primary key - either a single column or a
composite set of columns - for the endpoint.
}
\examples{
get_ok_pk(endpoint = "inventor")
get_ok_pk(endpoint = "patent/foreign_citation")

}
