% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patentsview-package.R
\docType{package}
\name{patentsview-package}
\alias{patentsview}
\alias{patentsview-package}
\title{patentsview: An R Client to the 'PatentsView' API}
\description{
Provides functions to simplify the 'PatentsView' API (\url{https://patentsview.org/apis/purpose}) query language, send GET and POST requests to the API's twenty seven endpoints, and parse the data that comes back.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/patentsview/index.html}
  \item Report bugs at \url{https://github.com/ropensci/patentsview/issues}
}

}
\author{
\strong{Maintainer}: Christopher Baker \email{chriscrewbaker@gmail.com}

Authors:
\itemize{
  \item Russ Allen \email{rrjallen@yahoo.com}
}

}
\keyword{internal}
