% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-dsl.R
\docType{data}
\name{qry_funs}
\alias{qry_funs}
\title{Query function list}
\format{An object of class \code{list} of length 14.}
\usage{
qry_funs
}
\value{
An object of class \code{pv_query}. This is basically just a simple list with a print method attached to it.
}
\description{
A list containing R functions that make it easy to write PatentsView queries. See the \href{https://github.com/crew102/patentsview/blob/master/vignettes/writing-queries.Rmd}{writing queries} vignette for details.
}
\examples{
qry_funs$eq(patent_date = "2001-01-01")

qry_funs$not(qry_funs$eq(patent_date = "2001-01-01"))

}
\keyword{datasets}

