% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-pv.R
\name{retrieve_linked_data}
\alias{retrieve_linked_data}
\title{Get HATEOS Data}
\usage{
retrieve_linked_data(url)
}
\arguments{
\item{url}{The link that was returned by the API on a previous call}
}
\value{
A character vector with field names, same as search_pv()
}
\description{
The new version of the api return HATEOAS style links to get more data
ex "inventor": "https://search.patentsview.org/api/v1/inventor/252373/"
They are new, Get only endpoints with no ability to pass f: s: or o: parameters
}
\examples{
\dontrun{

retrieve_linked_data("https://search.patentsview.org/api/v1/cpc_subgroup/G01S7:4811/")
}

}
