% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-pv.R
\name{retrieve_linked_data}
\alias{retrieve_linked_data}
\title{Get Linked Data}
\usage{
retrieve_linked_data(url, api_key = Sys.getenv("PATENTSVIEW_API_KEY"))
}
\arguments{
\item{url}{The link that was returned by the API on a previous call}

\item{api_key}{API key. See \href{https://patentsview.org/apis/keyrequest}{
Here} for info on creating a key.}
}
\value{
A character vector with field names, same as search_pv()
}
\description{
Some of the endpoints now returns HATEOAS style links to get more data.
ex "inventor": "https://search.patentsview.org/api/v1/inventor/252373/"
}
\examples{
\dontrun{

retrieve_linked_data(
  "https://search.patentsview.org/api/v1/cpc_subgroup/G01S7:4811/"
 )
}

}
