% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-pv.R
\name{search_pv}
\alias{search_pv}
\title{Search PatentsView data}
\usage{
search_pv(query, fields = NULL, endpoint = "patents", subent_cnts = FALSE,
  mtchd_subent_only = FALSE, page = 1, per_page = 25, all_pages = FALSE,
  sort = NULL, method = "GET", error_browser = getOption("browser"), ...)
}
\arguments{
\item{query}{The query you wish to send to the API endpoint. This paramater can be any one of the following:
\itemize{
 \item A character or JSON vector of length 1 with valid JSON query syntax (e.g., \code{'{"_gte":{"patent_date":"2007-01-04"}}'})
 \item A list which will be converted to JSON (e.g., \code{list("_gte" = list("patent_date" = "2007-01-04"))})
 \item A call to one of the query helper functions found in the \code{qry_funs} list (e.g., \code{with_qfuns(gte(patent_date = "2007-01-04"))}. See examples and/or the "writing queries" vignette.
}}

\item{fields}{A character vector of data fields that you want returned from the server. A value of \code{NULL} indicates that the default fields should be returned, as per the PatentsView API. Accecepible field values for a given endpoint can be found in the API's online documentation. For example, the field list for the patents endpoint can be found \href{http://www.patentsview.org/api/patent.html#field_list}{here}. Field lists for the six other endpoints are available on the PatentsView website as well...Note, you can use \code{\link{get_fields}} to get the field names for a given group, instead of typing out a bunch of field names.}

\item{endpoint}{A character vector of length 1 indicating which web service resource you wish to search. \code{endpoint} must be one of the following: \code{"patents", "inventors", "assignees", "locations", "cpc_subsections", "uspc_mainclasses", "nber_subcategories"}.}

\item{subent_cnts}{Do you want the total counts of unique subentities to be returned to you? This is equivlient to the \code{include_subentity_total_counts} parameter found \href{http://www.patentsview.org/api/query-language.html#options_parameter}{here}. See details for an explanation of subentities.}

\item{mtchd_subent_only}{Do you want all the subentities returned to you, even if they would be filtered by your query? This is the \code{matched_subentities_only} parameter found here: \href{http://www.patentsview.org/api/query-language.html#options_parameter}{here}. See details for an explanation of subentities.}

\item{page}{The page number of the results that you want returned.}

\item{per_page}{The number of records that should be returned per page. This value can be as high as 10,000 (e.g., \code{per_page = 10000}).}

\item{all_pages}{Do you want to download all possible pages of output? If \code{all_pages = TRUE}, the values of \code{page} and \code{per_page} are ignored.}

\item{sort}{A named character vector where the name indicates the field to sort by and the value indicates the direction of sorting ("asc" or "desc"). For example, \code{sort = c("patent_number" = "asc")} or \code{sort = c("patent_number" = "asc", "patent_date" = "desc")}. \code{sort = NULL} (the default) means do not sort the results.}

\item{method}{A character vector of length 1 indicating the HTTP method that you want to use to send the request. Possible values include \code{method = "GET"} or \code{method = "POST"}.}

\item{error_browser}{The program used to view any HTML error messages sent by the API. This should be either a character vector of length 1 giving the name of the program (assuming it is in your PATH...e.g., \code{error_browser = "chrome"}) or the full path to the program (e.g., \code{error_browser = "C:/Program Files (x86)/Google/Chrome/Application/chrome.exe"}) if it is not on your PATH. Alternatively, you can provide an R function to be called to invoke the error browser (e.g., \code{error_browser = rstudioapi::viewer}). Under Windows, \code{NULL} is also an allowed value for \code{error_browser} (and is the default), and implies that the file association mechanism will be used to determine which browser is used. To turn error browsing off when using Windows, set \code{error_browser = "false"}.}

\item{...}{Arguments passed along to httr's \code{\link[httr]{GET}} or \code{\link[httr]{POST}} function.}
}
\value{
A list with the following three elements:
\describe{
  \item{data_results}{A list with one element - a data frame containing the data returned by the server. Each row in the data frame corresponds to a unique value of the field denoted by the endpoint. For example, if you search the assignee endpoint, then the data frame will be on the assignee-level, where each row corresponds to a single assignee. Fields that are not on the assignee-level for this result set will be returned in a list column.}
  \item{query_results}{Entity counts across all pages of output (not just the page returned to you). If you set \code{subent_cnts = TRUE}, you will be returned both the counts of the main entities as well as any sub-entities that would be returned by your search. See description for a short explaination of entities/subentities.}
  \item{request}{Details of the HTTP request that was sent to the client. When you set \code{all_pages = TRUE}, \emph{you will only get a sample request. In other words, you will not be given multiple requests for the multiple calls made to the API to loop over the pages.}}
 }
}
\description{
Search PatentsView data
}
\details{
\strong{What is a subentity?}
}
\examples{

}

